requirements.txt
google-adk>=1.0.0
litellm>=1.30.0
google-cloud-aiplatform>=1.40.0
python-dotenv>=1.0.1
pydantic>=2.6.0
schedule>=1.2.1
